refactore into 2 tests files for feed an non feed stuff
